Revise the file lab-1i/fib.s to complete Fibonacci number function with non-recursive way.
Use for-loop to compute the fianl number step by step, which begins with number 0 and number 1.
Assign register R0 as output value when single loop finishs every time.
Assign register R3 as the second-to-laste output value.
Assign register R4 as the number of repetitions.
Assign register R5 as the temporary value.